au/iris-workers: Refactoring. Prepare to remove Aby3Store.storage
