+++ iris-mpc-cpu/src/execution/hawk_main/iris_worker.rs
@@ -1,0 +1,156 @@
+use crate::{
+    hawkers::shared_irises::SharedIrisesRef,
+    protocol::{ops::galois_ring_pairwise_distance, shared_iris::ArcIris},
+    shares::RingElement,
+};
+use core_affinity::CoreId;
+use crossbeam::channel::{Receiver, Sender};
+use iris_mpc_common::vector_id::VectorId;
+use itertools::Itertools;
+use std::{
+    cmp,
+    sync::{
+        atomic::{AtomicU64, Ordering},
+        Arc,
+    },
+};
+use tokio::sync::oneshot;
+
+#[derive(Debug)]
+enum IrisTask {
+    EvalDistanceBatch {
+        queries: Vec<ArcIris>,
+        vector_ids: Vec<VectorId>,
+        rsp: oneshot::Sender<Vec<RingElement<u16>>>,
+    },
+    RingPairwiseDistance {
+        input: Vec<Option<(ArcIris, ArcIris)>>,
+        rsp: oneshot::Sender<Vec<RingElement<u16>>>,
+    },
+}
+
+#[derive(Clone, Debug)]
+pub struct IrisPoolHandle {
+    workers: Arc<Vec<Sender<IrisTask>>>,
+    next_counter: Arc<AtomicU64>,
+    // TODO: Refactor such that this is not needed.
+    iris_store: SharedIrisesRef<ArcIris>,
+}
+
+impl IrisPoolHandle {
+    pub async fn eval_distance_batch(
+        &self,
+        queries: Vec<ArcIris>,
+        vector_ids: Vec<VectorId>,
+    ) -> Vec<RingElement<u16>> {
+        let (tx, rx) = oneshot::channel();
+        let task = IrisTask::EvalDistanceBatch {
+            queries,
+            vector_ids,
+            rsp: tx,
+        };
+        let _ = self.get_next_worker().send(task);
+        rx.await.unwrap()
+    }
+
+    pub async fn galois_ring_pairwise_distances(
+        &self,
+        input: Vec<Option<(ArcIris, ArcIris)>>,
+    ) -> Vec<RingElement<u16>> {
+        let (tx, rx) = oneshot::channel();
+        let task = IrisTask::RingPairwiseDistance { input, rsp: tx };
+        let _ = self.get_next_worker().send(task);
+        rx.await.unwrap()
+    }
+
+    fn get_next_worker(&self) -> &Sender<IrisTask> {
+        // fetch_add() wraps around on overflow
+        let idx = self.next_counter.fetch_add(1, Ordering::Relaxed) as usize;
+        let idx = idx % self.workers.len();
+        &self.workers[idx]
+    }
+}
+
+pub fn init_workers(
+    num_workers: usize,
+    shard_index: usize,
+    iris_store: SharedIrisesRef<ArcIris>,
+) -> IrisPoolHandle {
+    let core_ids = select_core_ids(shard_index);
+
+    // need to use at least one core
+    let cores_to_use = std::cmp::max(
+        1,
+        // minus one to leave core 0 alone.
+        // other stuff probably gets scheduled on core 0
+        std::cmp::min(num_workers, core_ids.len().saturating_sub(1)),
+    );
+
+    let mut channels = vec![];
+    for &core_id in core_ids.iter().take(cores_to_use) {
+        let (tx, rx) = crossbeam::channel::unbounded::<IrisTask>();
+        channels.push(tx);
+        let iris_store = iris_store.clone();
+        std::thread::spawn(move || {
+            let _ = core_affinity::set_for_current(core_id);
+            worker_thread(rx, iris_store);
+        });
+    }
+
+    IrisPoolHandle {
+        workers: Arc::new(channels),
+        next_counter: Arc::new(AtomicU64::new(0)),
+        iris_store,
+    }
+}
+
+fn worker_thread(ch: Receiver<IrisTask>, iris_store: SharedIrisesRef<ArcIris>) {
+    while let Ok(task) = ch.recv() {
+        match task {
+            IrisTask::EvalDistanceBatch {
+                queries,
+                vector_ids,
+                rsp,
+            } => {
+                let vectors = {
+                    let store = iris_store.data.blocking_read();
+
+                    vector_ids.iter().map(|v| store.get_vector(v)).collect_vec()
+                };
+
+                let pairs = queries
+                    .iter()
+                    .flat_map(|q| {
+                        vectors
+                            .iter()
+                            .map(|vector| vector.as_ref().map(|v| (q.clone(), v.clone())))
+                    })
+                    .collect::<Vec<_>>();
+
+                let r = galois_ring_pairwise_distance(pairs);
+                let _ = rsp.send(r);
+            }
+
+            IrisTask::RingPairwiseDistance { input, rsp } => {
+                let r = galois_ring_pairwise_distance(input);
+                let _ = rsp.send(r);
+            }
+        }
+    }
+}
+
+const SHARD_COUNT: usize = 2;
+
+pub fn select_core_ids(shard_index: usize) -> Vec<CoreId> {
+    let shard_index = shard_index % SHARD_COUNT;
+
+    let mut core_ids = core_affinity::get_core_ids().unwrap();
+    core_ids.reverse();
+    assert!(!core_ids.is_empty());
+
+    let shard_size = core_ids.len() / SHARD_COUNT;
+    let start = shard_index * shard_size;
+    let end = cmp::min(start + shard_size, core_ids.len());
+
+    core_ids[start..end].to_vec()
+}
